Upgrade the Closure Compiler: v20180805 -> v20180910
